cleanup cli
