+++ packages/cli/src/v2/index.js
@@ -1,15 +1,24 @@
-const init2 = require('./init2');
-const resolve2 = require('./resolve2');
-const verify2 = require('./verify2');
-const sign2 = require('./sign2');
-const encrypt2 = require('./encrypt2');
-const decrypt2 = require('./decrypt2');
+const init = require('./init');
+const resolve = require('./resolve');
+const verify = require('./verify');
+const sign = require('./sign');
+const encrypt = require('./encrypt');
+const decrypt = require('./decrypt');
+
+const commands = {
+  init,
+  resolve,
+  verify,
+  sign,
+  encrypt,
+  decrypt,
+};
+
+const register = (vorpal) => {
+  Object.keys(commands).map(command => commands[command](vorpal));
+};
 
 module.exports = {
-  init2,
-  resolve2,
-  verify2,
-  sign2,
-  encrypt2,
-  decrypt2,
+  ...commands,
+  register,
 };
