+++ packages/cli/src/v2/resolve.js
@@ -1,7 +1,7 @@
 const ghdid = require('@github-did/lib');
 
 module.exports = (vorpal) => {
-  vorpal.command('resolve2 <did>', 'resolve a github-did').action(async ({ did }) => {
+  vorpal.command('resolve <did>', 'resolve a github-did').action(async ({ did }) => {
     const didDocument = await ghdid.v2.func.resolver.resolve(did);
     // eslint-disable-next-line
     console.log(JSON.stringify(didDocument, null, 2));
