--- packages/cli/test.dec.json
@@ -1,4 +1,0 @@
-{
-  "@context": "https://w3id.org/identity/v1",
-  "givenName": "Alice"
-}
