--- packages/cli/test.enc.json
@@ -1,5 +1,0 @@
-{
-  "fromPublicKeyId": "did:github:OR13#kid=jNeDDagaBn466F-wH26YdQ5_NiabBvOlXTv5xItQakU",
-  "toPublicKeyId": "did:github:OR13#kid=jNeDDagaBn466F-wH26YdQ5_NiabBvOlXTv5xItQakU",
-  "cipherText": "-----BEGIN PGP MESSAGE-----\r\nVersion: OpenPGP.js v4.4.7\r\nComment: https://openpgpjs.org\r\n\r\nwX4DEeDfhKR9fhcSAgMEO4xasTJdslxXmItIpgssOlMkc5AoKw6yJkme5qag\r\n8fbE5cq1g4urp+JPscvAJDJtPXKYxpqOIjm+lImF67/EhzDptYJA/VzPYUoU\r\nhNxrxIz95omVQ3K6aYUL+TM2z0U1RtwOyfd5drBol1LcwGySqXLSwCYBdHHp\r\n5npeTh93H6AoYQgTiWm2D7AIJ29QK7GYrJJaKBRdOHXQnSwSVe61IQFI5M1g\r\nFgfJWsAfUrJFjWTguJCtLW2t8D7KrBkqbbwJfPVgEm5P1ffFZlodFGD6tNu9\r\npiq6pLL0VVHAqz+d2KM36iMMw+CJR/Sgl5M0GqU6kQIclE2TqrfNKqK63cHd\r\n1U29Gm0fA1OB8Tn/UqmK3lajePj4Mh+/HO/hrk+PV5+3dgrs04BnSHULxLRp\r\nPPJcsBx4mfaM37pV/utbuYYhQCLyvTbCSzL3Djh/9weq04lawcbfcPfxkoRw\r\n7Q==\r\n=VdPs\r\n-----END PGP MESSAGE-----\r\n"
-}
