--- packages/cli/test.signed.json
@@ -1,12 +1,0 @@
-{
-  "@context": "https://w3id.org/identity/v1",
-  "givenName": "Alice",
-  "proof": {
-    "type": "OpenPgpSignature2019",
-    "creator": "did:github:OR13#kid=jNeDDagaBn466F-wH26YdQ5_NiabBvOlXTv5xItQakU",
-    "domain": "GitHubDID",
-    "nonce": "77d17b4f4ab2a84db790b266f35dcb93",
-    "created": "2019-05-12T18:39:10.701Z",
-    "signatureValue": "-----BEGIN PGP SIGNATURE-----\r\nVersion: OpenPGP.js v4.4.7\r\nComment: https://openpgpjs.org\r\n\r\nwl4EARMIAAYFAlzYaE8ACgkQtAbTfc9Fw6k1nwD9HpqrP+gQOTi9oWFMIOwm\r\n6mI3U/GrRmFbNLqsyvQ7Z1YBAJf7OvFT3LE1NoMap5C1mkZ6jrkMWrW3KSVK\r\n8DSYe48s\r\n=BO1j\r\n-----END PGP SIGNATURE-----\r\n"
-  }
-}
